web/server: Add a simple, customizable HTTP server
The HTTP server may be customized by registering `Handlers`, which are
associated with given base paths/prefixes (e.g., "/gamepad" or "/img").

A service may implement the optional `LoopbackHandler` interface, for
services that must be able to access another local resource (e.g., if
a service must be able to read local files).