web/timer: Add a timer module
`timer` serves a timer that may be controlled remotelly, through JSON
encoded messages.

To modify the timer, send  POST request with the operation (one of
(`setup`, `start`, `stop`, `reset`, `add` or `sub`) and possibly an
associated time in millisecond (e.g., for the `setup` operation).

To retrieve the current time, issue a GET request to the service, which
will be returned in milliseconds.