cmd/mt-overlay: Mystery Tournament 15 overlay
Overlay to simplify restreaming Mystery Tournament 15 races. The overlay
must configured from a webpage (http://localhost:8088/config) and then it
may be accessed from one of following addresses:

* 1v1 overlay: http://localhost:8088/tmpl/index-1v1.go.html
* 2v2 overlay: http://localhost:8088/tmpl/index-2v2.go.html
* Automatic overlay: http://localhost:8088/tmpl/index.go.html

The last one checks for a flag in the server to decide whether to provide
the 1v1 or the 2v2 overlay. Meanwhile, the other addresses forces one race
overlay or the other.

The overlay also has a dashboard for controlling the timer and setting the
match's winner, which may be accessed from http://localhost:8088/dashboard.

Note that neither the images nor the fonts used by this overlay were
committed (intentionally)!