cmd/mfh-overlay: Add popup module
Add the endpoint `/mfh-handler/popup` to the server.

In pages, simply including the scripts `/script/popup.js` and
`/script/conn.js` and periodically calling `popup.update()` is enough
show/hide elements whenever the server signals that an element should be
changed.

To show a currently hidden element, call `popup.show()` the element's ID
and the timeout for hiding it again.

Elements should be hidden when they have a `hidden` class, and temporarily
removing that class from the element should be enough to make it visible
once again.