cmd/mfh-overlay: Add the race overlay
Add a configurable overlay, accessible in `/tmpl/overlay.go.html`, intended
to be used in restreams of 1v1 and 4-way races. Although it can seamlessly
change between the 2 layouts, by marking a option in the configuration
page, there are also two extra mappings for the overlay to automatically
access these modes:

* `/tmpl/1v1-overlay.go.html` - forcefully set the 1v1 layout
* `/tmpl/2v2-overlay.go.html` - forcefully set the 4-way layout

The overlay must be configured by accessing page `/res/config.html`, and it
may be controlled live (e.g., starting the timer or setting a racer as the
winner) by accessing page `/res/dashboard.html`.