cmd/mfh-overlay: Add overlay for Giant 2oopa
This overlay uses the main `tmpl/overlay.go.html` as its base (including it
as an `iframe`), but customizes it by adding images behind this base.

The growing Heavy Troopa is an animated `svg`, instead of an `img`, so it
may scale more smoothly and so the players' stream may be masked.