web/ram-store: Add a volatile storage module
`ram_store` stores arbitrary data for later retrieval. It's quite
similar to `tmpl`, however there's no processing of the received
objects, and any object type is accepted.

The accept methods are GET, DELETE, POST and PUT, and the last two are
handled identically. GET returns the same Content Type specified in the
POST/PUT request, and the URL identify the resource being accessed.