local/key-events: Add a local key logger
key_events listen to keyboard events and may both trigger configurable
events for specific keys or send this data to a `/ram_store`-like
endpoint, so it may later be used by another service. This module must
run locally, instead of being serviced as an endpoint.

The module may use two resources within a configurable base endpoint
(for example, '/ram_store/keyboard`):

    * base_endpoint + '/data': A 'application/json' encoded array in
which each index represents the state of the key (0 for released, 1 for
pressed);
    * base_endpoint + '/map': A 'application/json' encoded array naming
every key in the same order as they appear in base_endpoint + `/data`.

This base endpoint may be left unconfigured, in which case this module
will only trigger the registered key events.