cmd/keyboard-viewer: Add a simple keyboard viewer application
Add a simple keyboard viewer application that watches the keyboard and
reports it to a (locally running) remote service. They keyboard may be
viewed on http://localhost:8000.

If running on Windows, be sure to compile the key logging DLL (from
github.com/SirGFM/goLogKeys/winDll) and place it besides the executable.