gamepad-tools/monitor: Differentiate package and lib errors
Because of how the library errors were originally mapped into the package,
the addition of any new error to the library would either change the value
of package errors or would force the library to know about the package.

Space out the package errors so new library errors may be added without
conflicting with any package error.